feat: add affiliate crud, and add partner sign-up comission integration

+++ src/application/use-cases/partner/sign-up.ts
@@ -15,9 +15,13 @@ import type { AxiosError } from 'axios';
 import type { IOperatorRepository } from '../../contracts/repos/operator';
 import type { IQueueService } from '../../contracts/services/queue';
 import type { VerifyPartnerConfirmationDto } from './verify-confirmation';
+import type { IAffiliateRepository } from '../../contracts/repos/affiliate';
+import type { InviteToken } from '../../../domain/entities/invite-token';
+import type { Affiliate } from '../../../domain/entities/affiliate';
 
 export interface SignPartnerUpDto {
-  invite_token: string
+  invite_token?: string
+  referral_code?: string
   name: string
   company_name: string
   cpf: string
@@ -41,6 +45,7 @@ export class SignPartnerUpUseCase {
   constructor(
     private readonly partnerRepo: IPartnerRepository,
     private readonly inviteTokenRepo: IInviteTokenRepository,
+    private readonly affiliateRepo: IAffiliateRepository,
     private readonly treatmentRepo: ITreatmentRepository,
     private readonly walletRepo: IWalletRepository,
     private readonly operatorRepo: IOperatorRepository,
@@ -51,10 +56,19 @@ export class SignPartnerUpUseCase {
   ){}
 
   async execute(props: SignPartnerUpDto): Promise<Error | SignPartnerUpReturn> {
-    const inviteToken = await this.inviteTokenRepo.findByToken(props.invite_token)
+    let inviteToken: InviteToken | null = null
+    let affiliate: Affiliate | null = null
 
-    if (!inviteToken) {
-      return new Error("Token not found")
+    if (props.invite_token) {
+      inviteToken = await this.inviteTokenRepo.findByToken(props.invite_token)
+    }
+
+    if (props.referral_code) {
+      affiliate = await this.affiliateRepo.findByReferralCode(props.referral_code)
+    }
+
+    if (!inviteToken && !affiliate) {
+      return new Error("Invite token and affiliate not found")
     }
 
     const partnerAlreadyExists = await this.partnerRepo.findByCpf(props.cpf)
@@ -97,7 +111,8 @@ export class SignPartnerUpUseCase {
       lat,
       lng,
       treatments: treatments.filter(treatment => props.treatment_ids.includes(treatment.id)),
-      operator_id: inviteToken.operator_id
+      operator_id: inviteToken ? inviteToken?.operator_id : undefined,
+      affiliate_id: affiliate ? affiliate.id : undefined
     })
 
     const createWalletResult = await this.paymentService.createWallet({
@@ -140,20 +155,22 @@ export class SignPartnerUpUseCase {
       delay: 30 * 60 * 1000 // 30 min
     })
 
-    if (inviteToken.operator_id) {
-      const operator = await this.operatorRepo.findById(inviteToken.operator_id)
-      
-      if (operator && operator.sign_up_comission_percentage) {
-        const addTransactionToOperatorInviteResult = await this.inviteTokenRepo.addTransaction(inviteToken.id, createInitialPaymentResult.transaction_id)
+    if (inviteToken) {
+      if (inviteToken.operator_id) {
+        const operator = await this.operatorRepo.findById(inviteToken.operator_id)
         
-        if (!addTransactionToOperatorInviteResult) {
-          console.error("Error add transaction to invite token: ", addTransactionToOperatorInviteResult)
+        if (operator && operator.sign_up_comission_percentage) {
+          const addTransactionToOperatorInviteResult = await this.inviteTokenRepo.addTransaction(inviteToken.id, createInitialPaymentResult.transaction_id)
+          
+          if (!addTransactionToOperatorInviteResult) {
+            console.error("Error add transaction to invite token: ", addTransactionToOperatorInviteResult)
+          }
+        } else {
+          await this.inviteTokenRepo.delete(inviteToken.id)
         }
       } else {
         await this.inviteTokenRepo.delete(inviteToken.id)
       }
-    } else {
-      await this.inviteTokenRepo.delete(inviteToken.id)
     }
 
     return createInitialPaymentResult;
